moved editor
