+++ src/components/editor.tsx
